Submissions List , count, problem Delete

+++ frontend/src/components/ProblemTable.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useMemo } from 'react'
 import { useAuthStore } from '../store/useAuthStore'
-import { Link } from "react-router-dom"
-
+import { Link, useActionData } from "react-router-dom"
+import useAction from '../store/useAction'
 import { Bookmark, PencilIcon, Trash, TrashIcon, Plus } from "lucide-react";
 const ProblemTable = ({ problems }) => {
     const { authUser } = useAuthStore();
@@ -11,7 +11,9 @@ const ProblemTable = ({ problems }) => {
     const [selectedTag, setSelectedTag] = useState("ALL");
     const [currentPage, setCurrentPage] = useState(1);
 
-    const allTags = useMemo(() => {
+const {isDeletingProblem,onDeleteProblem} = useAction();
+
+const allTags = useMemo(() => {
         if (!Array.isArray(problems)) return [];
         const tagsSet = new Set();
         problems.forEach((p) => p.tags?.forEach((t) => tagsSet.add(t)));
@@ -47,7 +49,12 @@ const ProblemTable = ({ problems }) => {
     const difficulties = ["EASY", "MEDIUM", "HARD"];
 
     const handleDelete = (id) => {
-    onDeleteProblem(id);
+        const isConfirmed = confirm("Are you sure you want to delete this Problem ?")
+    if (isConfirmed) {
+          onDeleteProblem(id);
+    }
+
+  
   };
  const handleCreatePlaylist = async (data) => {
     await createPlaylist(data);
@@ -176,7 +183,15 @@ const ProblemTable = ({ problems }) => {
                                                                 onClick={() => handleDelete(problem.id)}
                                                                 className="btn btn-sm btn-error"
                                                             >
-                                                                <TrashIcon className="w-4 h-4 text-white" />
+
+                                                                {
+                                                                    isDeletingProblem ? (
+                                                                        <span className="loading loading-spinner text-white"></span>
+                                                                    ) : (
+                                                                         <TrashIcon className="w-4 h-4 text-white" />
+                                                                    )
+                                                                }
+                                                              
                                                             </button>
                                                             <button disabled className="btn btn-sm btn-warning">
                                                                 <PencilIcon className="w-4 h-4 text-white" />
